requirements.txt
alembic==1.15.2
bcrypt==4.3.0
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.2
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
SQLAlchemy==2.0.41
typing_extensions==4.13.2
Werkzeug==3.1.3
